+++ NEWS.rst
@@ -5,6 +5,7 @@ News
 Version 0.7.2 - 2016-..-..
 
   * NEW: added support for common DXF attribute line weight
+  * BUGFIX: MeshData.optimize() - corrected wrong vertex optimization
 
 Version 0.7.1 - 2016-02-21