+++ NEWS.rst
@@ -7,6 +7,7 @@ Version 0.7.2 - 2016-..-..
   * NEW: added support for common DXF attribute line weight
   * BUGFIX: MeshData.optimize() - corrected wrong vertex optimization
   * BUGFIX: can open DXF files without existing layout management table
+  * BUGFIX: restore module structure ezdxf.const
 
 Version 0.7.1 - 2016-02-21